技术上：重构工程，将三种投资品进行抽象，界面工程只有一个，再加3个配置。handler中操作表格方法抽象，仅保留取数代码。减少魔法值字符串使用。引入线程池，避免随时创建线程。 功能上：基金和股票的默认cron表达式为A股开始时间。表达式时间重叠的情况加入了锁机制，避免无意义的压力。 界面上：表格默认自动缩放。
